requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.3
devtools==0.5.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
mypy==0.761
mypy-extensions==0.4.3
Pillow==7.0.0
pydantic==1.4
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
wrapt==1.11.2
yarl==1.4.2
